refactor: cookies utils
